requirements.txt
# FastAPI and ASGI server
fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
gunicorn>=21.2.0,<23.0.0

# Database
motor>=3.3.2,<4.0.0
pymongo>=4.6.1,<5.0.0
dnspython>=2.5.0,<3.0.0

# Authentication
python-jose[cryptography]>=3.3.0,<4.0.0

# Environment & Config
python-dotenv>=1.0.0,<2.0.0

# HTTP clients
httpx>=0.26.0,<1.0.0
requests>=2.31.0,<3.0.0

# Data validation
pydantic>=2.5.3,<3.0.0
google-auth>=2.25.2,<3.0.0
